requirements.txt
absl-py==0.10.0
aiohttp==3.6.2
antlr4-python3-runtime==4.8
async-timeout==3.0.1
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
dataclasses==0.7
future==0.18.2
google-auth==1.22.0
google-auth-oauthlib==0.4.1
grpcio==1.32.0
hydra-core==1.0.3
idna==2.10
idna-ssl==1.1.0
importlib-metadata==2.0.0
importlib-resources==3.0.0
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.2
multidict==4.7.6
numpy==1.19.2
oauthlib==3.1.0
omegaconf==2.0.2
opencv-python==4.4.0.44
packaging==20.4
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-lightning==0.9.0
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
tensorboard==2.2.0
tensorboard-plugin-wit==1.7.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.50.0
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==1.0.1
yarl==1.6.0
zipp==3.2.0
